% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check convergence of REMA model}
\usage{
check_convergence(mod, ret = FALSE, f = "")
}
\arguments{
\item{mod}{output from \code{\link{fit_rema}}}

\item{ret}{T/F, return list? Default = FALSE, just prints to console}
}
\value{
a list with at least the first three of these components: \describe{
  \item{\code{$convergence}}{From \code{\link[stats:nlminb]{stats::nlminb}},
  "0 indicates successful convergence for nlminb"} \item{\code{$maxgr}}{Max
  absolute gradient value, from `max(abs(mod$gr(mod$opt$par)))`}
  \item{\code{$maxgr_par}}{Name of parameter with max gradient}
  \item{\code{$is_sdrep}}{If \code{\link[TMB:sdreport]{TMB::sdreport}} was
  performed for this model, this indicates whether it performed without
  error} \item{\code{$na_sdrep}}{If \code{\link[TMB:sdreport]{TMB::sdreport}}
  was performed without error for this model, this indicates which (if any)
  components of the diagonal of the inverted hessian were returned as NA} }
}
\description{
Access quick convergence checks from `TMB` and `nlminb`. Function modified
from
\href{https://github.com/timjmiller/wham/blob/3a056359121bc1a911ed6a95c9203db4db456baa/R/check_convergence.R}{wham::check_convergence}.
}
\examples{
\dontrun{
# placeholder for example
}
}
\seealso{
\code{\link{fit_rema}}, \code{\link{fit_tmb}},
  \code{\link[stats:nlminb]{stats::nlminb}}
}
