% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_rema_models.R
\name{compare_rema_models}
\alias{compare_rema_models}
\title{Plot REMA model comparisons and return AIC values when appropriate}
\usage{
compare_rema_models(
  rema_models,
  save = FALSE,
  filetype = "png",
  path = NULL,
  xlab = NULL,
  biomass_ylab = "Biomass",
  cpue_ylab = "CPUE"
)
}
\arguments{
\item{rema_models}{list of REMA models to be compared. Each REMA model in the
list should be a list object output from \code{\link{fit_rema}}}

\item{save}{(optional) logical (T/F) save figures as \code{filetype} in
\code{path}. Default = FALSE. NOT YET IMPLEMENTED.}

\item{filetype}{(optional) character string; type of figure file. Default =
'png'. NOT YET IMPLEMENTED.}

\item{path}{(optional) directory path to location where figure files are to
be saved if \code{save = TRUE}. NOT YET IMPLEMENTED.}

\item{xlab}{(optional) label for x-axis of biomass and CPUE plots (e.g.
'Year'). Default = NULL.}

\item{biomass_ylab}{(optional) label for y-axis of biomass plots (e.g.
'Biomass (t)'). Default = 'Biomass'.}

\item{cpue_ylab}{(optional) label for y-axis of CPUE plots (e.g. 'Relative
Population Number'). Default = 'CPUE'.}
}
\value{
a list with the following items:
  \describe{
    \item{\code{$output}}{A list of tidied dataframes that include parameter
    estimates, biomass and optional CPUE data, and REMA model predictions for
    each model to be compared. Results for a given variable are only included
    if they are applicable to all comparison models. For example, if CPUE is
    fit in one model but not another, \code{compare$output$cpue_by_strata)}
    will return an informational message instead of a dataframe. See
    \code{\link{tidy_rema}} for more information.}
    \item{\code{$plots}}{ggplot2 figure objects of \code{compare$output} data.}
    \item{\code{$aic}}{A dataframe of Akaike Information Criteria (AIC)
    values. Only output if the underlying models are fit to the same data.}
}
}
\description{
Takes list of REMA models from from \code{\link{fit_rema}}, and returns a
list of \code{ggplot2} objects to be plotted or saved, a list of
\code{\link{tidy_rema}} data.frames, and AIC values.
}
\examples{
\dontrun{
# placeholder for example
}
}
\seealso{
\code{\link{tidy_rema}}, \code{\link{plot_rema}}
}
