% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_sdrep.R
\name{do_sdrep}
\alias{do_sdrep}
\title{Get sdreport() from TMB model}
\usage{
do_sdrep(model, save.sdrep = TRUE)
}
\arguments{
\item{model}{Output from an optimized TMB model
\code{\link[stats:nlminb]{stats::nlminb}}}
}
\value{
the sdreport from \code{\link[stats:nlminb]{TMB::sdreport}}
}
\description{
Function originally from:
https://github.com/timjmiller/wham/blob/97577f1d619607ec40fdb8d3097f30668af16999/R/fit_wham.R#L186
}
