% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rema.R
\name{fit_rema}
\alias{fit_rema}
\title{Fit REMA model}
\usage{
fit_rema(
  input,
  n.newton = 0,
  do.sdrep = TRUE,
  model = NULL,
  do.check = FALSE,
  MakeADFun.silent = TRUE,
  do.fit = TRUE,
  save.sdrep = TRUE
)
}
\arguments{
\item{input}{Named list output from \code{\link{prepare_rema_input}}, which
includes the following components needed to fit model using
\code{\link[TMB:MakeADFun]{TMB::MakeADFun}}:
\describe{
  \item{\code{$data}}{Data, a list of data objects for model fitting or
  specification (e.g., user-defined pentalties, index pointers, etc.). A
  required input to \code{\link[TMB]{MakeADFun}}.}
  \item{\code{$par}}{Parameters, a list of all random and fixed effects
  parameter objects.  A required input to \code{\link[TMB]{MakeADFun}}.}
  \item{\code{$map}}{Map, a mechanism for collecting and fixing parameters
  in TMB.  An input to \code{\link[TMB]{MakeADFun}}.}
  \item{\code{$random}}{Character vector defining the parameters to treat
  as random effects. An input to \code{\link[TMB]{MakeADFun}}.}
  \item{\code{$model_name}}{Character, name of the model, e.g. \code{"GOA
  shortraker with LLS by depth strata"}. Useful for model comparison.}
}}

\item{n.newton}{integer, number of additional Newton steps after
optimization. Not an option that is currently needed, but is passed to
\code{\link{fit_tmb}}. Default = \code{0}.}

\item{do.sdrep}{T/F, calculate standard deviations of model parameters? See
\code{\link[TMB]{sdreport}}. Default = \code{TRUE}.}

\item{model}{(optional), a previously fit rema model.}

\item{do.check}{T/F, check if model parameters are identifiable? Passed to
\code{\link{fit_tmb}}. Runs internal function \code{check_estimability},
originally provided by https://github.com/kaskr/TMB_contrib_R/TMBhelper.
Default = \code{TRUE}.}

\item{MakeADFun.silent}{T/F, Passed to silent argument of
\code{\link[TMB:MakeADFun]{TMB::MakeADFun}}. Default = \code{TRUE}.}

\item{do.fit}{T/F, fit the model using \code{fit_tmb}. Default = \code{TRUE}.}

\item{save.sdrep}{T/F, save the full \code{\link[TMB]{TMB::sdreport}} object?
If \code{FALSE}, only save
\code{\link[TMB:summary.sdreport]{summary.sdreport}} to reduce model object
file size. Default = \code{TRUE}.}
}
\value{
a fit TMB model with additional output if specified:
  \describe{
    \item{\code{$rep}}{List of derived quantity estimates (e.g. estimated
    biomass)}
    \item{\code{$sdrep}}{Parameter estimates (and standard errors if
    \code{do.sdrep = TRUE})}
  }
}
\description{
Fits the compiled REMA model using
\code{\link[TMB:MakeADFun]{TMB::MakeADFun}} and
\code{\link[stats:nlminb]{stats::nlminb}}. Source code and documentation modified from
\href{https://github.com/timjmiller/wham/blob/master/R/fit_wham.R}{wham::fit_wham}.
}
\details{
Future development: Implement one-step-ahead (OSA) residuals for evaluating
model goodness-of-fit \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}).
OSA residuals are more appropriate than standard residuals for models with
random effects
(\href{https://link.springer.com/article/10.1007/s10651-017-0372-4}{Thygeson
et al. (2017)}. See
\href{https://github.com/timjmiller/wham/blob/master/R/fit_wham.R}{wham} for
an example of OSA implementation and additional OSA residual options (e.g.
full Gaussian approximation instead of the (default) generic method using
\code{osa.opts=list(method="fullGaussian")}.
}
\examples{
\dontrun{
# place holder for example code
}
}
\seealso{
\code{\link{fit_tmb}},
         \code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}
}
