% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osa_residuals.R
\name{get_osa_residuals}
\alias{get_osa_residuals}
\title{Get one-step-head (OSA)}
\usage{
get_osa_residuals(
  rema_model,
  options = list(method = "fullGaussian", parallel = TRUE)
)
}
\arguments{
\item{rema_model}{list out output from \code{\link{fit_rema}}, which includes
model results but also inputs. Of note to OSA residual calculations is the
\code{rema_model$input$osa} object, which is a data.frame containing all
the data or observations fit in the model that will have a residuals
associated with them.}

\item{options}{list of options for calculating OSA residuals, passed to
\code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}. Default:
\code{options = list(method = "fullGaussian", parallel = TRUE)}.
Alternative methods include "cdf", "oneStepGeneric",
"oneStepGaussianOffMode", and "oneStepGaussian". Users implementing the
Tweedie distribution should use "cdf" methods.}
}
\value{
a list of tidied data.frames containing the biomass and CPUE survey
  residuals with accompanying data, as well as a QQ-plot, histogram
  of residuals, and plots of residuals~year and residuals~fitted values by
  strata for the biomass and CPUE survey.
}
\description{
Takes the rema model output from \code{\link{fit_rema}} and returns OSA
residuals calculated using
\code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}} with accompanying
residual analysis plots. IMPORTANT: users implementing the Tweedie
distribution should use "cdf" methods. See `options` for more details.
}
\examples{
\dontrun{
# placeholder for example
}
}
\seealso{
\code{\link{tidy_rema}}
}
