% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rema_generics.R
\name{plot.rema}
\alias{plot.rema}
\title{Plots the distribution of the test statistic results from a \code{rema} object}
\usage{
\method{plot}{rema}(x, ...)
}
\arguments{
\item{x}{a \code{rema} object holding distribution results from the
\code{rema} function}

\item{...}{other arguments that can be specified in plot}
}
\value{
A barplot of the distribution of the test statistic.
}
\description{
Plots the distribution of the test statistic results from a \code{rema} object
with the probabilities on the y-axis and the test statistics on the x-axis.
}
\details{
The observed value of the test statistic is marked with an asterisk. The dark
gray bars mark the probabilities that contribute to the two-sided p-value.
The light gray bars highlight probabilities that are greater than the
probability of the observed test statistic and consequently do not contribute
to the two-sided p-value. If \code{dist} is set to \code{FALSE} in the
\code{rema} object, the plot function will return an error.
}
