% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_re_dat.R
\name{read_re_dat}
\alias{read_re_dat}
\title{Convert ADMB version of the RE model output to REMA inputs}
\usage{
read_re_dat(filename, biomass_strata_names = NULL, cpue_strata_names = NULL)
}
\arguments{
\item{filename}{name of ADMB output file to be read (e.g. rwout.rep)}

\item{biomass_strata_names}{(optional) a vector of character names
corresponding to the names of the biomass survey strata. Vector should be
in the same order as the columns of srv_est in rwout.rep}

\item{cpue_strata_names}{(optional) a vector of character names corresponding
to the names of the CPUE survey strata. Vector should be in the same order
as the columns of srv_est_LL in rwout.rep in the version of the ADMB RE
model that accepts an additional survey index}
}
\value{
object of type "list" with biomass optional cpue survey data in
  long format, and initial parameter values for log_biomass_pred (the random
  effects matrix), ready for input into REMA
}
\description{
Read the report file from the ADMB version of the RE model (rwout.rep) and
convert it into long format survey data estimates with CVs for input into
REMA.
}
