% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rep.R
\name{read_rep}
\alias{read_rep}
\title{Read ADMB .rep file and return an R object of type 'list'}
\usage{
read_rep(fn)
}
\arguments{
\item{fn}{full path and name of ADMB output file to be read}
}
\value{
object of type "list" with ADMB outputs therein
}
\description{
Code modified from original function provided by Steve Martell, D'Arcy N. Webber
called by \code{\link{read_re_dat}}
}
\examples{
\dontrun{
read_rep(fn = 'inst/example_data/goasr.rep')
}
}
