% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_extra_cv.R
\name{tidy_extra_cv}
\alias{tidy_extra_cv}
\title{Tidy estimates of extra biomass or CPUE index CV}
\usage{
tidy_extra_cv(tidy_rema, save = FALSE, path = NULL, alpha_ci = 0.05)
}
\arguments{
\item{tidy_rema}{list out output from \code{\link{tidy_rema}}, which includes
model results but also inputs}

\item{save}{(optional) logical (T/F) save figures as \code{filetype} in
\code{path}. Default = FALSE. NOT YET IMPLEMENTED.}

\item{path}{(optional) directory path to location where figure files are to
be saved if \code{save = TRUE}. NOT YET IMPLEMENTED.}

\item{alpha_ci}{(optional) the significance level for generating confidence
intervals. Default = 0.05}
}
\value{
a list with the following items:
  \describe{
    \item{\code{$parameter_estimates}}{A data.frame of fixed effects
    parameters in REMA (e.g. log_PE and log_q) with standard errors and
    confidence intervals that have been transformed from log space to natural
    space for ease of interpretation. }
    \item{\code{$biomass_by_strata}}{A tidy, long format data.frame of model
    predicted and observed biomass by biomass survey strata. This data.frame
    is now enhanced with new columns that include log-space standard error
    and associated confidence intervals that account for additional estimated
    observation error.}
    \item{\code{$cpue_by_strata}}{A tidy, long format data.frame of model
    predicted and observed CPUE by CPUE survey strata. This data.frame is now
    enhanced with new columns that include log-space standard error and
    associated confidence intervals that account for additional estimated
    observation error. If REMA is not run in multi-survey mode, or if CPUE
    data are not provided, an explanatory character string with instructions
    for fitting to CPUE data is returned.}
    \item{\code{$biomass_by_cpue_strata}}{A tidy, long format data.frame of
    model predicted biomass by CPUE survey strata. Note that observed/summed
    biomass observations are not returned in case there are missing values in
    one stratum but not another within a given year. This output is reserved
    for instances when the number of biomass strata exceeds that of CPUE
    survey strata, but the user wants to visualize predicted biomass at the
    same resolution as the CPUE predictions. In other scenarios, a character
    string is returned explaining the special use case for this object.}
    \item{\code{$total_predicted_biomass}}{A tidy, long format data.frame of
    total model predicted biomass summed across all biomass survey strata. If
    only one stratum is used (i.e. the univariate RE), the predicted values
    will be the same as \code{output$biomass_by_strata}.}
    \item{\code{$total_predicted_cpue}}{A tidy, long format data.frame of
    total model predicted CPUE summed across all CPUE survey strata. If only
    one stratum is used (i.e. the univariate RE), the predicted values will
    be the same as \code{output$cpue_by_strata}. If The CPUE survey index
    provided was defined as not summable in prepare_rema_input(), an
    character string will be returned explaining how to change this using the
    'sum_cpue_index' in \code{?prepare_rema_input} if appropriate.}
  }
}
\description{
Takes list output from \code{\link{tidy_rema}} and returns the same list with
enhanced versions of the \code{biomass_by_strata} and \code{cpue_by_strata}
when appropriate. These enhanced dataframes include three new columns,
\code{tot_log_obs_cv}, \code{tot_obs_lci}, and \code{tot_obs_uci}, which
represent combined log-space standard error and associated confidence
intervals that include both assumed and estimated additional observation
error.
}
\examples{
\dontrun{
# placeholder for example
}
}
\seealso{
\code{\link{tidy_rema}}
}
