% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_osa_residuals.R
\name{get_osa_residuals}
\alias{get_osa_residuals}
\title{Get one-step-head (OSA)}
\usage{
get_osa_residuals(
  rema_model,
  options = list(method = "fullGaussian", parallel = F)
)
}
\arguments{
\item{rema_model}{list out output from \code{\link{fit_rema}}, which includes
model results but also inputs. Of note to OSA residual calculations is the
\code{rema_model$input$osa} object, which is a data.frame containing all
the data or observations fit in the model that will have a residuals
associated with them.}

\item{options}{list of options for calculating OSA residuals, passed to
\code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}. Default:
\code{options = list(method = "cdf", parallel = TRUE)}.}
}
\value{
a list of tidied data.frames containing the biomass and CPUE survey
  residuals with accompanying data, as well as ggplot2 plots to visual
  results from the OSA residual analysis.
}
\description{
Takes list output from \code{\link{tidy_rema}} and returns a list of
\code{ggplot2} objects to be plotted or saved. This feature is
experimental, and OSA residuals are still in Beta mode in
\code{\link[TMB:oneStepPredict]{TMB::oneStepPredict}}. The default "cdf"
method sometimes results in NA values for residuals, especially when
observation errors are small.
}
\examples{
\dontrun{
# placeholder for example
}
}
\seealso{
\code{\link{tidy_rema}}
}
